refactor : embedded redis 사용 방식 수정
